Add a landmark (only) particle filter SLAM.
This SLAM solution is based on *known* landmarks (or features).
Each landmarks position is updated using a Kalman Filter.

It supports two types of measurement models, a polar (or range-bearing)
and a Cartesian model.

WARNING: This solutions has *not* been tested yet. It is a work-in-progress
that needs to be tested.