[slam] Switch to 6 DoF landmark state. Improve Hybrid SLAM.
We now track the 3D pose of each landmark.

In the Hybrid Particle Filter SLAM, the weights of the landmarks are
kept separated from the laser weights. The laser completely dominates
the error and the same gain cannot be applied to both weights.
Therefore, different gains are used. After applying the gain the weights
are added.