[slam] Add support for multimode.
A particle filter is capable of representing several modes of a
distribution but we were using only the best particle (or sample).
Now, the particle filter calculates the best mode and provides its mean,
that is, the pose of the robot is a mean value of all samples in the
mode cluster weighted by the normalized weight.