New global localization strategy.
The global localization now spreads the particles around a landmark
(if present) or around the free space of the map and lets the particle
filter evolve.

This requires somes changes to how the SLAM system works on mapping is
paused. Now, when mapping is paused each particle will use the same map
for localization, which is the proper way to do it. The maps are keept
and reused when mapping is resumed.